pyproject.toml
[tool.poetry]
name = "fairplay"
version = "0.1.0"
description = "Using GenAI to forge fairer societal agreements."
authors = ["Jenny Tai <jenny.tai@uwaterloo.ca>", "Sergio M. Ferro <sm.ferro54@gmail.com>"]
packages = [
    {include = "fairplay", from = "src"}
]
license = "MIT License"
readme = "README.md"

[tool.poetry.scripts]
fairplay = "fairplay.main:main"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
annotated-types = "0.7.0"
anyio = "4.6.0"
autogen = "0.3.0"
certifi = "2024.8.30"
charset-normalizer = "3.3.2"
diskcache = "5.6.3"
distro = "1.9.0"
docker = "7.1.0"
flaml = {extras = ["automl"], version = "^2.3.2"}
h11 = "0.14.0"
httpcore = "1.0.6"
httpx = "0.27.2"
idna = "3.10"
jiter = "0.5.0"
langchain = ">=0.3.15,<0.4.0"
langchain-openai = ">=0.3.2,<0.4.0"
numpy = "1.26.4"
openai = ">=1.51.0"
packaging = "24.1"
pydantic = ">=2.10.6,<3.0.0"
pydantic-core = ">=2.23.4"
python-dotenv = ">=1.0.1,<2.0.0"
regex = "2024.9.11"
requests = "2.32.3"
sniffio = "1.3.1"
termcolor = "2.4.0"
tiktoken = "0.7.0"
tqdm = "4.66.5"
typing-extensions = "4.12.2"
urllib3 = "2.2.3"
graphviz = "^0.20.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[tool.setuptools.packages.find]
where = ["src/"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]

build-backend = "poetry.core.masonry.api"
